# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WalkMission API",
    "description": "오하꼼(WalkMission) 백엔드 API 명세서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "JWT Auth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "tags": [
          "User & Auth Domain"
        ],
        "summary": "회원가입",
        "description": "새로운 사용자를 등록합니다.",
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "가입 완료",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "요청 오류 (code로 원인 구분: INVALID_INPUT, TOKEN_EXPIRED, MISSION_NOT_FOUND 등)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "서버 오류 (INTERNAL_ERROR, PLACE_SEARCH_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reissue": {
      "post": {
        "tags": [
          "User & Auth Domain"
        ],
        "summary": "토큰 재발급",
        "description": "리프레시 토큰을 검증하고 새로운 액세스 토큰과 리프레시 토큰을 발급합니다.",
        "operationId": "reissue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReissueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReissueResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "요청 오류 (code로 원인 구분: INVALID_INPUT, TOKEN_EXPIRED, MISSION_NOT_FOUND 등)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "서버 오류 (INTERNAL_ERROR, PLACE_SEARCH_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "User & Auth Domain"
        ],
        "summary": "로그인 및 토큰 발급",
        "description": "사용자 로그인을 처리하고 액세스 토큰을 발급합니다.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "요청 오류 (code로 원인 구분: INVALID_INPUT, TOKEN_EXPIRED, MISSION_NOT_FOUND 등)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "서버 오류 (INTERNAL_ERROR, PLACE_SEARCH_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SignupRequest": {
        "required": [
          "birth",
          "email",
          "nickname",
          "password",
          "regionCode"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "password123!"
          },
          "nickname": {
            "type": "string",
            "description": "닉네임",
            "example": "walking_master"
          },
          "birth": {
            "type": "string",
            "description": "생년월일",
            "format": "date",
            "example": "1990-01-01"
          },
          "regionCode": {
            "type": "string",
            "description": "동네 코드 (법정동/행정동)",
            "example": "1168010100"
          }
        },
        "description": "회원가입 요청 DTO"
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "userUuid": {
            "type": "string",
            "description": "사용자 고유 UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nickname": {
            "type": "string",
            "description": "닉네임",
            "example": "walking_master"
          },
          "message": {
            "type": "string",
            "description": "응답 메시지",
            "example": "회원가입이 완료되었습니다."
          }
        },
        "description": "회원가입 응답 DTO"
      },
      "ReissueRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "리프레시 토큰",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "description": "토큰 재발급 요청 DTO"
      },
      "ReissueResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "새 액세스 토큰",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "새 리프레시 토큰",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "description": "토큰 재발급 응답 DTO"
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "password123!"
          }
        },
        "description": "로그인 요청 DTO"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "액세스 토큰",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "리프레시 토큰",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "userUuid": {
            "type": "string",
            "description": "사용자 고유 UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nickname": {
            "type": "string",
            "description": "닉네임",
            "example": "walking_master"
          }
        },
        "description": "로그인 응답 DTO"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP 상태 코드",
            "format": "int32",
            "example": 400
          },
          "code": {
            "type": "string",
            "description": "에러 코드 (프론트 분기용)",
            "example": "NOT_ENOUGH_DISTANCE"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 보여줄 수 있는 메시지",
            "example": "목적지 반경 50m 이내에 도착하지 않았습니다."
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "추가 정보 (없으면 생략). 예: 남은 거리, 필드별 검증 오류",
              "example": {
                "currentDistanceMeters": 120
              }
            },
            "description": "추가 정보 (없으면 생략). 예: 남은 거리, 필드별 검증 오류",
            "example": {
              "currentDistanceMeters": 120
            }
          }
        },
        "description": "공통 에러 응답"
      }
    },
    "securitySchemes": {
      "JWT Auth": {
        "type": "http",
        "name": "JWT Auth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}